powerbi: refactoring the embedded sdk into a versioned folder
