+++ internal/services/powerbi/sdk/2021-01-01/capacities/method_suspend_autorest.go
