+++ internal/services/powerbi/sdk/2021-01-01/capacities/model_dedicatedcapacities.go
